// src/interfaces/ItemInputErrorInterface.ts
interface ItemInputErrorInterface {
    item_category: boolean;
    item_category_msg: string;
    item_type: boolean;
    item_type_msg: string;
    item_code: boolean;
    item_code_msg: string;
    item_desc: boolean;
    item_desc_msg: string;
    item: boolean;
    item_msg: string;
    schedule_class: boolean;
    schedule_class_msg: string;
    alloy: boolean;
    alloy_msg: string;
    size: boolean;
    size_msg: string;
    weight: boolean;
    weight_msg: string;
    cut_weight: boolean;
    cut_weight_msg: string;
    cut_length: boolean;
    cut_length_msg: string;
    cut_width: boolean;
    cut_width_msg: string;
    std_material_used: boolean;
    std_material_used_msg: string;
    finished_code: boolean;
    finished_code_msg: string;
    finished_desc: boolean;
    finished_desc_msg: string;
}
export default ItemInputErrorInterface;